pyproject.toml
[project]
name = "mcp-cli-client"
version = "0.1.0"
description = "MCP CLI Client with multi-LLM support"
dependencies = [
    "typer>=0.16.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "requests>=2.32.0",
    "mcp>=1.0.0",
]

[project.scripts]
mcp-client = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"